// src/app/models/masters-mpc/packagingpresentation.ts
export class Packagingpresentation {
    id: number = -1;
    name: string = "";
    idPackingType: number = -1;
    packingType: string = "";
    createdByUser: string = "";
    createdByUserId: number = -1;
    updatedByUser: string = "";
    createdDate: Date;
    updatedDate: Date;
    active: boolean = false;
    initialSetup: boolean = false;
}